CHANGE: Journey times (including waiting and transfer times) are now represented as 32-bit rather than 16-bit integers. This allows (in theory) for journey times of over 800 years (in game time) rather than the previous maximum of 109 hours (approx. 4.5 days). This will be important for long sea voyages, especially when portals are introduced. FIX: Journey times of over 109 hours were not properly treated. FIX: Crash in (rare) cases when an intermediate stop becomes the nearest stop to the ultimate destination during a journey FIX: Passengers were recorded incorrectly as unhappy or too slow instead of no route in at least some cases ADD: New town name ending NOTE: The above change increments the saved game version.

+++ simconvoi.h
@@ -51,7 +51,7 @@ class replace_data_t;
 * The table of point-to-point average journey times.
 * @author jamespetts
 */
-typedef koordhashtable_tpl<id_pair, average_tpl<uint16> > journey_times_map;
+typedef koordhashtable_tpl<id_pair, average_tpl<uint32> > journey_times_map;
 
 # define entry x
 # define reversed y
